# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/testnet",
    "paths": {
        "/mainnet/GetBalance/{qaddress}": {
            "get": {
                "description": "GetBalance gRPC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mainnet"
                ],
                "summary": "GetBalance for mainnet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QRL address",
                        "name": "qaddress",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/testnet/GetBalance/{qaddress}": {
            "get": {
                "description": "GetBalance gRPC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "testnet"
                ],
                "summary": "GetBalance for testnet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QRL address",
                        "name": "qaddress",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}